Unify the WPF and Cocoa implementations of the PreviewFactoryService
